Add CSS to intro-to-css 02
Create a class named odd to be applied to odd numbered elements.
Create an id named second to be applied to the second element.
Create a class named third to be applied to the third element in addition
to the applied odd class.
Create an id named fourth to be applied to the fourth element.
Each class and id has the appropriate CSS property.